+++ recipes/install-zip.rb
@@ -3,7 +3,7 @@
 # Cookbook:: webpi
 # Recipe:: install-zip
 #
-# Copyright:: 2011-2016, Chef Software, Inc.
+# Copyright:: 2011-2017, Chef Software, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
